src/main/webapp/app/entities/study-academic-year/delete/study-academic-year-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { StudyAcademicYearService } from '../service/study-academic-year.service';

import { StudyAcademicYearDeleteDialogComponent } from './study-academic-year-delete-dialog.component';

describe('StudyAcademicYear Management Delete Component', () => {
  let comp: StudyAcademicYearDeleteDialogComponent;
  let fixture: ComponentFixture<StudyAcademicYearDeleteDialogComponent>;
  let service: StudyAcademicYearService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [StudyAcademicYearDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(StudyAcademicYearDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(StudyAcademicYearDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(StudyAcademicYearService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
